requirements.txt
colpali-engine>=0.3.0,<0.4.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
pillow>=9.0.0
# Added for GOT-OCR-2.0
safetensors>=0.3.1
